impr: RELENG-7322 docker-build: add git repo labels
Add the labels 'git.commit-sha' and 'git.repository' to images built
using the docker-build workflow, which values are the git repository
name (e.g. "scality/cloudserver") and the git commit SHA used to build
the image.

The immediate need is for Integration to tell which commit hash to
pull Cloudserver tests from (INTGR-907).